bug fix: instance. refactoring bug :  int scale -> float scale,  SetGraphicsRootShaderResourceView  코드위치 mQueue.WaitUntil(cmdAllocs

+++ 3D/3D/GameObject.h
@@ -13,7 +13,7 @@ public:
 class Transform
 {
 public:
-	int scale = 1;
+	float scale = 1;
 	Vector3 position;
 };
